+++ services/docker/general.go
@@ -20,6 +20,7 @@ var containerPorts = map[string]string{
 	"nginx":   "8082",
 	"solr":    "8983",
 	"redis":   "6379",
+	"adminer": "8080",
 	"unison":  "5000",
 }