+++ services/docker/generate.go
@@ -120,6 +120,22 @@ func UnmarshalledDefaults() conf.ComposeDotTok {
 		}
 	}
 
+	if conf.GetConfig().Services.Mailhog.Enabled {
+		v := "v1.0.0"
+		errMailhog := yaml.Unmarshal(dockertmpl.EnableMailhog(v), &tokStruct)
+		if errMailhog != nil {
+			log.Fatalf("Error enabling Mailhog in Compose file: %v", err)
+		}
+	}
+
+	if conf.GetConfig().Services.Adminer.Enabled {
+		v := "4-standalone"
+		errAdminer := yaml.Unmarshal(dockertmpl.EnableAdminer(v), &tokStruct)
+		if errAdminer != nil {
+			log.Fatalf("Error enabling Adminer in Compose file: %v", err)
+		}
+	}
+
 	if conf.GetConfig().Services.Memcache.Enabled {
 		var v string
 		if conf.GetConfig().Tokaido.Betacontainers {
@@ -165,6 +181,8 @@ func getCustomTok() []byte {
 	dc.Services.Memcache.Enabled = false
 	dc.Services.Solr.Enabled = false
 	dc.Services.Redis.Enabled = false
+	dc.Services.Mailhog.Enabled = false
+	dc.Services.Adminer.Enabled = false
 
 	cc, err := yaml.Marshal(dc)
 	if err != nil {