+++ services/docker/templates/compose.go
@@ -57,6 +57,25 @@ func EnableRedis(version string) []byte {
       - "6379"`)
 }
 
+// EnableMailhog ...
+func EnableMailhog(version string) []byte {
+	return []byte(`services:
+  mailhog:
+    image: mailhog/mailhog:` + version + `
+    ports:
+      - "1025"
+      - "8025"`)
+}
+
+// EnableAdminer ...
+func EnableAdminer(version string) []byte {
+	return []byte(`services:
+  adminer:
+    image: adminer:` + version + `
+    ports:
+      - "8080"`)
+}
+
 // SetUnisonVersion ...
 func SetUnisonVersion(version string) []byte {
 	return []byte(`services: